- big refactor of code and routes. View Excel Sheet for more information on REST API

+++ src/main/java/com/example/MonaServer/Helper/UsernameXPoints.java
@@ -1,3 +1,3 @@
 package com.example.MonaServer.Helper;
 
-public record UsernameXPoints (String username, int points) {}
+public record UsernameXPoints (String username, Long points) {}
